Refactored top level app component into its own file
