+++ client/src/constants.js
@@ -1,0 +1,3 @@
+export const Types = {
+  CARD: 'card'
+};
